up & feat
- injeted auth user in user action store
- company controller userStore method has been moved to user controller inot a generic method called modelStore
- abort added on attach/detach in order to inform the user
- coordinates field request modified to 'nullable' during tests
- added space between ui form fields
- search form extracted into a component called 'search-box'
- attach/create user modal prop title
- route moved for modelStore in user group
- sqlite company test moved under Company directory
- removed randomly problematic test, as this actions are covered in request tests
- added request tests for attaching / creating new company members

+++ app/Http/Controllers/UserController.php
@@ -213,7 +213,7 @@ class UserController extends Controller
 
         abort_unless(App::make($policy)->edit(Auth::user(), $model), 401);
         $this->authorize('edit', $user);
-
+        abort_unless(!$model->users()->find($user), 404);
         $model->users()->attach($user);
 
         return back()
@@ -236,7 +236,7 @@ class UserController extends Controller
 
         abort_unless(App::make($policy)->edit(Auth::user(), $model), 401);
         $this->authorize('edit', $user);
-
+        abort_unless($model->users()->find($user), 404);
         $model->users()->detach($user);
 
         return back()
@@ -245,6 +245,29 @@ class UserController extends Controller
                 'User unlinked',
                 'User has been unlinked from ' . $modelName
             ));
+    }
+
+    public function modelStore(StoreUserRequest $request, string|int $modelId, UserStoreAction $action): RedirectResponse
+    {
+        $modelName = Collection::make($request->route()->parameters())->keys()->first();
+        $model = RelatedModel::from($modelName)->entity($modelId);
+        $policy = RelatedModel::from($modelName)->policy();
+
+        abort_unless(App::make($policy)->edit(Auth::user(), $model), 401);
+
+        $attributes = $request->safe()->all();
+        $attributes['active'] = $request->boolean('active');
+        $user = $action->handle($attributes);
+
+        $model->users()->attach($user);
+        Auth::user()->peerAttach($user);
+
+        return back()
+          ->with('message', self::responseMessage(
+              'success',
+              'User created & linked',
+              'User has been created and linked to ' . $modelName,
+          ));
     }
 
     // ADMIN
